% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remindStart-package.R
\docType{package}
\name{remindStart-package}
\alias{remindStart}
\alias{remindStart-package}
\title{remindStart: Start the REMIND Integrated Assessment Model}
\description{
Start the REMIND Integrated Assessment Model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://johanneskoch94.github.io/remindStart/}
}

}
\author{
\strong{Maintainer}: Johannes Koch \email{jokoch@pik-potsdam.de}

}
\keyword{internal}
