% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{start}
\alias{start}
\title{start}
\usage{
start(remind = ".", configFile = NULL, restart = FALSE, testOneRegi = FALSE)
}
\arguments{
\item{remind}{A character vector with the path to the remind main-folder. By default, it is assumed that the working
directory is the remind main-folder.}

\item{configFile}{NULL or a character vector with the path to a scenario config csv file.}

\item{restart}{TRUE or FALSE}

\item{testOneRegi}{TRUE or FALSE}
}
\description{
Start the REMIND model.
}
\examples{
\dontrun{
# Start default remind run
start()

# Start multiple runs with a scenario config file
start(configFile = "config/my_config.csv")

# Restart a run
start(restart = TRUE)

# Start from the command line
Rscript -e "remindStart::start()" --args config/my_config.csv
}
}
